requirements.txt
openai
pandas
python-dotenv
numpy
tqdm
streamlit
